+++ packages/parcel-plugin-codegen/src/index.ts
@@ -1,0 +1,5 @@
+import type * as ParcelBundler from 'parcel-bundler';
+
+export = function (bundler: ParcelBundler) {
+  bundler.addAssetType('codegen', require.resolve('./CodeGenAsset'));
+};