+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 0.15.0 (tbd)
 
 - Moved package to dedicated repository
+- Updated `parcel-plugin-at-alias` to use TypeScript
+- Updated `parcel-plugin-codegen` to use TypeScript